requirements.txt
streamlit-image-select
open_clip_torch